refactor: estrutura de pastas do projeto
