+++ .gitignore
@@ -138,5 +138,3 @@ cython_debug/
 
 # VS Code settings
 .vscode/
-
-chroma_store
